refactor(frontend): rename MaterialIcon to Icon
