+++ frontend/src/components/NavbarItem.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import MaterialIcon from "@/components/MaterialIcon.vue";
+import Icon from "@/components/common/Icon.vue";
 
 const { title, icon, href } = defineProps<{ title: string; icon: string; href: string }>();
 </script>
@@ -10,7 +10,7 @@ const { title, icon, href } = defineProps<{ title: string; icon: string; href: s
     class="flex flex-none flex-col items-center justify-center desktop:size-navbar-width desktop:gap-2 desktop:transition-colors
       desktop:hoctive:bg-current/5 mobile:mx-auto mobile:size-navbar-height mobile:gap-1"
   >
-    <MaterialIcon :type="icon" class="desktop:text-7 mobile:text-6" />
+    <Icon :type="icon" class="desktop:text-7 mobile:text-6" />
     <div class="select-none text-xs/none" v-text="title"></div>
   </RouterLink>
 </template>
