+++ frontend/src/components/TimelineItemGroup.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import Button from "@/components/common/Button.vue";
 import Collapse from "@/components/common/Collapse.vue";
+import Icon from "@/components/common/Icon.vue";
 import List from "@/components/common/List.vue";
-import MaterialIcon from "@/components/MaterialIcon.vue";
 import TimelineItem from "@/components/TimelineItem.vue";
 import { type MailboxEvent } from "@/types/MailboxEvent";
 import { ref } from "vue";
@@ -18,7 +18,7 @@ const collapsed = ref<boolean>(false);
         mobile:gap-2 mobile:py-2"
       @click="collapsed = !collapsed"
     >
-      <MaterialIcon
+      <Icon
         class="text-6 text-gray-500 transition-transform"
         type="keyboard_arrow_down"
         :class="{ '-rotate-90': collapsed }"
