+++ frontend/src/components/common/Icon.vue
